[Small Refactor] Moving the inference engine def outside the inference config
